Refactor web app.js into composed frontend modules
