+++ hydra_detect/web/static/js/ui/toast.js
@@ -1,0 +1,49 @@
+'use strict';
+
+window.HydraModules = window.HydraModules || {};
+
+window.HydraModules.createToastService = function createToastService() {
+    const toasts = [];
+    const MAX_TOASTS = 3;
+    const TOAST_DEDUP_MS = 5000;
+
+    function dismissToast(el) {
+        el.classList.add('dismissing');
+        setTimeout(() => {
+            el.remove();
+            const idx = toasts.findIndex(t => t.el === el);
+            if (idx !== -1) toasts.splice(idx, 1);
+        }, 200);
+    }
+
+    function showToast(message, type = 'error') {
+        const container = document.getElementById('toast-container');
+        if (!container) return;
+
+        const now = Date.now();
+        const isDupe = toasts.some(t => t.message === message && (now - t.time) < TOAST_DEDUP_MS);
+        if (isDupe) return;
+
+        while (toasts.length >= MAX_TOASTS) {
+            const oldest = toasts.shift();
+            if (oldest.el && oldest.el.parentNode) {
+                oldest.el.classList.add('dismissing');
+                setTimeout(() => oldest.el.remove(), 200);
+            }
+        }
+
+        const el = document.createElement('div');
+        el.className = `toast toast-${type}`;
+        el.textContent = message;
+        el.addEventListener('click', () => dismissToast(el));
+        container.appendChild(el);
+
+        toasts.push({ el, message, time: now });
+        setTimeout(() => dismissToast(el), 10000);
+    }
+
+    return {
+        showToast,
+        dismissToast,
+    };
+};
