+++ hydra_detect/web/templates/base.html
@@ -154,7 +154,15 @@
 
     <!-- ── Scripts ── -->
     <script src="/static/js/vendor/Sortable.min.js"></script>
-    <script src="/static/js/app.js"></script>
+    <script src="/static/js/state/store.js"></script>
+    <script src="/static/js/router/view-router.js"></script>
+    <script src="/static/js/polling/poller-manager.js"></script>
+    <script src="/static/js/stream/stream-controller.js"></script>
+    <script src="/static/js/ui/toast.js"></script>
+    <script src="/static/js/ui/modal.js"></script>
+    <script src="/static/js/api/client.js"></script>
+    <script src="/static/js/preflight/preflight.js"></script>
+    <script src="/static/js/main.js"></script>
     <script src="/static/js/panels.js"></script>
     <script src="/static/js/ops.js"></script>
     <script src="/static/js/config.js"></script>
